add rds_json output filter
