+++ config
@@ -5,11 +5,13 @@ HTTP_MODULES="$HTTP_MODULES \
 
 NGX_ADDON_SRCS="$NGX_ADDON_SRCS \
 		$ngx_addon_dir/src/ngx_http_mysql_module.c \
+		$ngx_addon_dir/src/ngx_http_mysql_handler.c \
 		$ngx_addon_dir/src/ngx_http_mysql_output.c"
 
 NGX_ADDON_DEPS="$NGX_ADDON_DEPS \
 		$ngx_addon_dir/src/ngx_http_mysql_module.h \
 		$ngx_addon_dir/src/ngx_http_mysql_output.h \
+		$ngx_addon_dir/src/ngx_http_mysql_handler.h \
 		$ngx_addon_dir/src/ngx_http_mysql_ddebug.h \
 		$ngx_addon_dir/src/resty_dbd_stream.h"
