+++ src/ngx_http_mysql_module.c
@@ -5,7 +5,10 @@
   > Created Time: 2012年12月21日 星期五 15时46分57秒
  ************************************************************************/
 
-#include <ngx_http_mysql_module.h>
+#include "ngx_http_mysql_module.h"
+#include "ngx_http_mysql_output.h"
+#include "ngx_http_mysql_ddebug.h"
+#include "ngx_http_mysql_handler.h"
 
 
 static ngx_command_t ngx_http_mysql_commands[] = {
@@ -141,566 +144,7 @@ ngx_module_t ngx_http_mysql_module = {
 };
 
 
-ngx_int_t ngx_http_mysql_transaction_handler(ngx_http_request_t *r){
-
-	ngx_http_mysql_loc_conf_t *mslcf;
-	ngx_http_mysql_srv_conf_t *msscf;
-	ngx_str_t query;
-	ngx_str_t trans_lev = ngx_string("SET SESSION TRANSACTION ISOLATION LEVEL READ COMMITTED");
-	ngx_str_t rollback = ngx_string("ROLLBACK");
-	MYSQL mysql, *sock;
-	ngx_chain_t *out;
-	ngx_http_mysql_node_t *mnode;
-	ngx_int_t ret;
-	ngx_http_mysql_ctx_t *ctx;
-	ngx_uint_t i;
-	ngx_http_mysql_trans_loc_conf_t * transaction_sqls;
-
-	msscf = ngx_http_get_module_srv_conf(r, ngx_http_mysql_module);
-
-	mslcf = ngx_http_get_module_loc_conf(r, ngx_http_mysql_module);
-
-	ctx = ngx_http_get_module_ctx(r, ngx_http_mysql_module);
-
-	if (ctx == NULL) {
-
-		ctx = ngx_pcalloc(r->connection->pool, sizeof(ngx_http_mysql_ctx_t));
-
-		ngx_http_set_ctx(r, ctx, ngx_http_mysql_module);
-	}
-
-	sock = NULL;
-	mnode = NULL;
-	out = NULL;
-	ret = NGX_ERROR;
-
-	if (msscf->max_conn != NGX_CONF_UNSET 
-			&& msscf->max_conn) 
-	{
-		/* take connection from pool */
-
-		if (msscf->free_node == NULL) {
-
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"not enough MySQL connections: %d", msscf->max_conn);
-
-			goto quit;
-		}
-
-		mnode = msscf->free_node;
-
-		msscf->free_node = msscf->free_node->next;
-
-		sock = &mnode->mysql;
-
-	} else {
-
-		/* connection-per-request */
-
-		sock = &mysql;
-	}
-
-	if (mnode == NULL || !mnode->ready) {
-
-		ngx_log_error(NGX_LOG_INFO, r->connection->log, 0, 
-				"connecting to MySQL");
-
-		mysql_init(sock);
-
-		if (!(sock = mysql_real_connect(sock,
-						NGXCSTR(msscf->host), 
-						NGXCSTR(msscf->user), 
-						NGXCSTR(msscf->password),
-						NGXCSTR(msscf->database), 
-						msscf->port == NGX_CONF_UNSET ? 0 : msscf->port, 
-						NULL,
-						msscf->multi == 1 ? CLIENT_MULTI_STATEMENTS : 0))) 
-		{
-
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"couldn't connect to MySQL engine: %s",
-					mysql_error(sock));
-
-			goto quit;
-		}
-
-		/*set charset*/
-		if (msscf->charset.len 
-				&& mysql_set_character_set(sock, NGXCSTR(msscf->charset)))
-		{
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"error setting MySQL charset: %s",
-					mysql_error(sock));
-
-			goto quit;
-		}
-
-		sock->reconnect = 1;
-
-		if (mnode)
-			mnode->ready = 1;
-
-		ngx_log_error(NGX_LOG_DEBUG, r->connection->log, 0, "mysql_auto_commit(%d)", msscf->mysql_auto_commit);
-
-		/*set auto_commit*/
-		if (0 != mysql_autocommit(sock, msscf->mysql_auto_commit))
-		{
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"error setting MySQL auto_commit: %s",
-					mysql_error(sock));
-
-			goto quit;
-		}
-
-		/* set TRANSACTION_REPEATABLE_READ to trans level*/
-		if (0 != mysql_real_query(sock, NGXCSTR(trans_lev), trans_lev.len))
-		{
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"error setting MySQL transaction level: %s",
-					mysql_error(sock));
-
-			goto quit;
-		}
-	}
-
-
-	/* set current MySQL connection for escapes */
-
-	ctx->current = sock;
-
-
-	transaction_sqls = mslcf->transaction_sqls;
-
-	ngx_log_error(NGX_LOG_DEBUG, r->connection->log, 0, "query len(%d)", transaction_sqls->len);
-
-	for (i = 0; i < transaction_sqls->len-1; ++i) 
-	{
-
-		ngx_log_error(NGX_LOG_DEBUG, r->connection->log, 0, "transaction_sql sql[%d]:%V", i, &transaction_sqls[i].sql); 
-
-		/* not found script $arg_id in sql */
-		if (NULL == transaction_sqls[i].query_lengths && NULL == transaction_sqls[i].query_values)
-		{
-			query = transaction_sqls[i].sql;
-		}
-		else
-		{
-			if (ngx_http_script_run(r, &query, transaction_sqls[i].query_lengths->elts, 0, transaction_sqls[i].query_values->elts) == NULL)
-			{
-				ctx->current = NULL;
-
-				goto quit;
-			}
-
-		}
-
-		ctx->current = NULL;
-
-		ngx_log_error(NGX_LOG_DEBUG, r->connection->log, 0, "transaction_sql query[%d]:%V", i, &query); 
-
-		if (mysql_real_query(sock, NGXCSTR(query), query.len)) 
-		{
-
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"MySQL read_query failed (%s)",
-					mysql_error(sock));
-			
-			if (mysql_real_query(sock, NGXCSTR(rollback), rollback.len))
-			{
-				ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-						"MySQL read_query failed (%s)",
-						mysql_error(sock));
-			}
-
-			goto quit;
-		}
-	}
-
-
-	if (out == NULL) {
-
-		/* no result */
-
-		out = (ngx_chain_t*)ngx_palloc(r->connection->pool, sizeof(ngx_chain_t));
-		out->next = NULL;
-		out->buf = ngx_create_temp_buf(r->connection->pool, 1);
-		*out->buf->last++ = 'y';
-		*out->buf->last++ = 'e';
-		*out->buf->last++ = 's';
-		out->buf->last_buf = 1;
-	}
-	r->headers_out.status = NGX_HTTP_OK;
-
-	if (!r->subrequest_in_memory) {
-
-		ngx_http_send_header(r);
-
-		ngx_http_output_filter(r, out);
-	}
-
-	ret = NGX_OK;
-
-
-quit:
-
-	if (mnode != NULL) {
-
-		mnode->next = msscf->free_node;
-		msscf->free_node = mnode;
-	}
-	else if (sock != NULL) {
-
-		mysql_close(sock);
-	}
-	if (r->subrequest_in_memory) {
-
-		ctx->subreq_out = out;
-	}
-
-	return ret;
-}
-
-ngx_int_t ngx_http_mysql_handler(ngx_http_request_t *r) {
-
-	/* NB:
-	   Remember this handler is completely ASYNCHRONOUS 
-	   when used with nginx-mtask-module;
-	   Blocking calls are intercepted to switch user-space
-	   context to NGINX event cycle.
-	 */
-
-	ngx_http_mysql_loc_conf_t *mslcf;
-	ngx_http_mysql_srv_conf_t *msscf;
-	ngx_str_t query;
-	MYSQL mysql, *sock;
-	MYSQL_RES *res;
-	MYSQL_ROW row;
-	char *value;
-	size_t len;
-	int n, num_fields, status;
-	ngx_chain_t *out, *node, *prev;
-	uint64_t auto_id;
-	ngx_http_mysql_node_t *mnode;
-	ngx_int_t ret;
-	ngx_http_mysql_ctx_t *ctx;
-
-	msscf = ngx_http_get_module_srv_conf(r, ngx_http_mysql_module);
-	
-	mslcf = ngx_http_get_module_loc_conf(r, ngx_http_mysql_module);
-
-	ctx = ngx_http_get_module_ctx(r, ngx_http_mysql_module);
-
-	if (ctx == NULL) {
-	
-		ctx = ngx_pcalloc(r->connection->pool, sizeof(ngx_http_mysql_ctx_t));
-
-		ngx_http_set_ctx(r, ctx, ngx_http_mysql_module);
-	}
-
-	sock = NULL;
-	mnode = NULL;
-	out = NULL;
-	ret = NGX_ERROR;
-
-	if (msscf->max_conn != NGX_CONF_UNSET 
-			&& msscf->max_conn) 
-	{
-		/* take connection from pool */
-
-		if (msscf->free_node == NULL) {
-
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"not enough MySQL connections: %d", msscf->max_conn);
-
-			goto quit;
-		}
-
-		mnode = msscf->free_node;
-
-		msscf->free_node = msscf->free_node->next;
-
-		sock = &mnode->mysql;
-
-	} else {
-
-		/* connection-per-request */
-
-		sock = &mysql;
-	}
-
-	if (mnode == NULL || !mnode->ready) {
-
-		ngx_log_error(NGX_LOG_INFO, r->connection->log, 0, 
-				"connecting to MySQL");
-
-		mysql_init(sock);
-
-		if (!(sock = mysql_real_connect(sock,
-						NGXCSTR(msscf->host), 
-						NGXCSTR(msscf->user), 
-						NGXCSTR(msscf->password),
-						NGXCSTR(msscf->database), 
-						msscf->port == NGX_CONF_UNSET ? 0 : msscf->port, 
-						NULL,
-						msscf->multi == 1 ? CLIENT_MULTI_STATEMENTS : 0))) 
-		{
-
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"couldn't connect to MySQL engine: %s",
-					mysql_error(sock));
-
-			goto quit;
-		}
-
-		if (msscf->charset.len 
-				&& mysql_set_character_set(sock, NGXCSTR(msscf->charset)))
-		{
-			ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-					"error setting MySQL charset: %s",
-					mysql_error(sock));
-
-			goto quit;
-		}
-
-		sock->reconnect = 1;
-
-		if (mnode)
-			mnode->ready = 1;
-	}
-
-	/* set current MySQL connection for escapes */
-
-	ctx->current = sock;
-
-	if (ngx_http_script_run(r, &query, mslcf->query_lengths->elts, 0,
-				mslcf->query_values->elts) == NULL)
-	{
-		ctx->current = NULL;
-
-		goto quit;
-	}
-
-	ctx->current = NULL;
-
-	if (mysql_query(sock, NGXCSTR(query))) {
-
-		ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-				"MySQL query failed (%s)",
-				mysql_error(sock));
-
-		goto quit;
-	}
-
-	out = NULL;
-
-	prev = NULL;
-
-	do {
-
-		if (!mysql_field_count(sock)) {
-
-			/* no data returned 
-			   check if there's auto-id pending */
-
-			auto_id = mysql_insert_id(sock);
-
-			if (auto_id) {
-
-				ngx_log_debug1(NGX_LOG_DEBUG_HTTP, r->connection->log, 0, 
-						"MySQL auto-inserted id=%uL", auto_id);
-
-				node = (ngx_chain_t*)ngx_palloc(r->connection->pool, sizeof(ngx_chain_t));
-				node->next = NULL;
-
-				node->buf = ngx_create_temp_buf(r->connection->pool, 32);
-				node->buf->last = ngx_slprintf(node->buf->pos, node->buf->end, "%uL\n", auto_id);
-
-				if (prev)
-					prev->next = node;
-
-				else
-					out = node;
-
-				prev = node;
-
-			}
-
-		} else {
-
-			if (!(res = mysql_store_result(sock))) {
-
-				ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-						"couldn't get MySQL result: %s",
-						mysql_error(sock));
-
-				goto quit;
-			}
-
-			num_fields = mysql_num_fields(res);
-
-			while((row = mysql_fetch_row(res))) {
-
-				for(n = 0; n < num_fields; ++n) {
-
-					value = row[n];
-
-					ngx_log_debug1(NGX_LOG_DEBUG_HTTP, r->connection->log, 0, 
-							"MySQL value returned '%s'", value ? value : "NULL");
-
-					len = value ? strlen(value) : 0;
-
-					node = (ngx_chain_t*)ngx_palloc(r->connection->pool, sizeof(ngx_chain_t));
-					node->next = NULL;
-					node->buf = ngx_create_temp_buf(r->connection->pool, len + 1);
-
-					if (value)
-						memcpy(node->buf->pos, value, len);
-
-					node->buf->pos[len] = '\n';
-					node->buf->last += ++len;
-
-					if (prev)
-						prev->next = node;
-
-					else
-						out = node;
-
-					prev = node;
-				}
-			}
-
-			mysql_free_result(res);
-		}
-
-		status = -1;
-
-		if (msscf->multi == 1) {
-
-			status = mysql_next_result(sock);
-
-			if (status > 0) {
-
-				ngx_log_error(NGX_LOG_ALERT, r->connection->log, 0, 
-						"couldn't get next MySQL result: %s",
-						mysql_error(sock));
-
-				goto quit;
-			}
-
-		}
-
-	} while(!status);
-
-	if (prev)
-		prev->buf->last_buf = 1;
-
-	if (out == NULL) {
-
-		/* no result */
-
-		out = (ngx_chain_t*)ngx_palloc(r->connection->pool, sizeof(ngx_chain_t));
-		out->next = NULL;
-		out->buf = ngx_create_temp_buf(r->connection->pool, 1);
-		*out->buf->last++ = '\n';
-		out->buf->last_buf = 1;
-	}
-
-	r->headers_out.status = NGX_HTTP_OK;
-
-	if (!r->subrequest_in_memory) {
-		
-		ngx_http_send_header(r);
-
-		ngx_http_output_filter(r, out);
-	}
-
-	ret = NGX_OK;
-
-quit:
-
-	if (mnode != NULL) {
-
-		mnode->next = msscf->free_node;
-		msscf->free_node = mnode;
-
-	} else if (sock != NULL)
-		mysql_close(sock);
-
-	if (r->subrequest_in_memory) {
-
-		ctx->subreq_out = out;
-	}
-
-	return ret;
-}
-
-static ngx_int_t ngx_http_mysql_subrequest_handler(ngx_http_request_t *r)
-{
-	ngx_http_mysql_loc_conf_t *mslcf;
-	ngx_http_request_t *sr;
-	ngx_http_mysql_ctx_t *ctx;
-	u_char *q;
-	ngx_str_t uri;
-	ngx_str_t args = ngx_null_string;
-	
-	mslcf = ngx_http_get_module_loc_conf(r, ngx_http_mysql_module);
-
-	ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0, 
-			"mysql subrequest handler");
-
-	if (mslcf->subreq_values == NULL
-			|| mslcf->subreq_lengths == NULL)
-	{
-		return NGX_DECLINED;
-	}
-
-	ctx = ngx_http_get_module_ctx(r, ngx_http_mysql_module);
-
-	if (ctx != NULL)
-		return NGX_DECLINED;
-
-	if (ngx_http_script_run(r, &uri, mslcf->subreq_lengths->elts, 0,
-				mslcf->subreq_values->elts) == NULL)
-	{
-		return NGX_ERROR;
-	}
-
-	ngx_log_debug1(NGX_LOG_DEBUG_HTTP, r->connection->log, 0, 
-			"mysql subrequest handler '%V'", &uri);
-
-	q = ngx_strlchr(uri.data, uri.data + uri.len, '?');
-
-	if (q != NULL) {
-		args.data = q + 1;
-		args.len = uri.data + uri.len - q - 1;
-		uri.len = q - uri.data;
-	}
-
-	if (ngx_http_subrequest(r, 
-			&uri,
-			&args,
-			&sr,
-			NULL,
-			NGX_HTTP_SUBREQUEST_WAITED|NGX_HTTP_SUBREQUEST_IN_MEMORY) != NGX_OK)
-	{
-		return NGX_ERROR;
-	}
-
-	if (sr == NULL)
-		return NGX_ERROR;
-
-	/* set the same context to parent & subrequest;
-	   result will go there */
-
-	ctx = ngx_pcalloc(r->connection->pool, sizeof(ngx_http_mysql_ctx_t));
-
-	ngx_http_set_ctx(r, ctx, ngx_http_mysql_module);
-	ngx_http_set_ctx(sr, ctx, ngx_http_mysql_module);
-
-	return NGX_DONE;
-}
-
-static ngx_int_t ngx_http_mysql_get_subrequest_variable(ngx_http_request_t *r,
+ngx_int_t ngx_http_mysql_get_subrequest_variable(ngx_http_request_t *r,
 		    ngx_http_variable_value_t *v, uintptr_t data)
 {
 	ngx_http_mysql_ctx_t *ctx;
@@ -747,7 +191,7 @@ static ngx_int_t ngx_http_mysql_get_subrequest_variable(ngx_http_request_t *r,
 	return NGX_OK;
 }
 
-static ngx_int_t ngx_http_mysql_get_escaped_variable(ngx_http_request_t *r,
+ngx_int_t ngx_http_mysql_get_escaped_variable(ngx_http_request_t *r,
 		    ngx_http_variable_value_t *v, uintptr_t data)
 {
 	ngx_http_mysql_ctx_t *ctx;
@@ -782,7 +226,7 @@ static ngx_int_t ngx_http_mysql_get_escaped_variable(ngx_http_request_t *r,
 	return NGX_OK;
 }
 
-static void* ngx_http_mysql_create_srv_conf(ngx_conf_t *cf)
+void* ngx_http_mysql_create_srv_conf(ngx_conf_t *cf)
 {
 	ngx_http_mysql_srv_conf_t *conf = ngx_pcalloc(cf->pool, sizeof(ngx_http_mysql_srv_conf_t));
 	
@@ -797,7 +241,7 @@ static void* ngx_http_mysql_create_srv_conf(ngx_conf_t *cf)
 	return conf;
 }
 
-static char* ngx_http_mysql_merge_srv_conf(ngx_conf_t *cf, void *parent, void *child)
+char* ngx_http_mysql_merge_srv_conf(ngx_conf_t *cf, void *parent, void *child)
 {
 	ngx_http_mysql_srv_conf_t *prev = parent;
 	ngx_http_mysql_srv_conf_t *conf = child;
@@ -843,19 +287,33 @@ static char* ngx_http_mysql_merge_srv_conf(ngx_conf_t *cf, void *parent, void *c
 	return NGX_CONF_OK;
 }
 
-static void* ngx_http_mysql_create_loc_conf(ngx_conf_t *cf)
+void* ngx_http_mysql_create_loc_conf(ngx_conf_t *cf)
 {
 	ngx_http_mysql_loc_conf_t *conf = ngx_pcalloc(cf->pool, sizeof(ngx_http_mysql_loc_conf_t));
+
+	conf->output_handler = NGX_CONF_UNSET_PTR;
 	
 	return conf;
 }
 
-static char* ngx_http_mysql_merge_loc_conf(ngx_conf_t *cf, void *parent, void *child)
+char* ngx_http_mysql_merge_loc_conf(ngx_conf_t *cf, void *parent, void *child)
 {
 	ngx_http_mysql_loc_conf_t *prev = parent;
 	ngx_http_mysql_loc_conf_t *conf = child;
 
 	ngx_log_debug0(NGX_LOG_INFO, cf->log, 0, "mysql merge loc");
+
+
+
+	if (conf->output_handler == NGX_CONF_UNSET_PTR) {
+		if (prev->output_handler == NGX_CONF_UNSET_PTR) {
+			/* default */
+			conf->output_handler = ngx_mysql_output_rds;
+		} else {
+			/* merge */
+			conf->output_handler = prev->output_handler;
+		}
+	}
 
 
 	if (conf->query_lengths == NULL)
@@ -880,7 +338,7 @@ static char* ngx_http_mysql_merge_loc_conf(ngx_conf_t *cf, void *parent, void *c
 	return NGX_CONF_OK;
 }
 
-static char* ngx_http_mysql_query(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
+char* ngx_http_mysql_query(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
 {
 	ngx_http_mysql_loc_conf_t *mslcf = conf;
 	ngx_http_mtask_loc_conf_t *mlcf;
@@ -912,7 +370,7 @@ static char* ngx_http_mysql_query(ngx_conf_t *cf, ngx_command_t *cmd, void *conf
 	return NGX_CONF_OK;
 }
 
-static char* ngx_http_mysql_transaction(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
+char* ngx_http_mysql_transaction(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
 {
 	ngx_http_mysql_loc_conf_t *mslcf = conf;
 	ngx_http_mtask_loc_conf_t *mlcf;
@@ -978,7 +436,7 @@ static char* ngx_http_mysql_transaction(ngx_conf_t *cf, ngx_command_t *cmd, void
 	return NGX_CONF_OK;
 }
 
-static char* ngx_http_mysql_subrequest(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
+char* ngx_http_mysql_subrequest(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
 {
 	ngx_http_mysql_loc_conf_t *mslcf = conf;
 	ngx_str_t *value;
@@ -1026,7 +484,7 @@ static char* ngx_http_mysql_subrequest(ngx_conf_t *cf, ngx_command_t *cmd, void 
 	return NGX_CONF_OK;
 }
 
-static char* ngx_http_mysql_escape(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
+char* ngx_http_mysql_escape(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
 {
 	ngx_str_t *value;
 	ngx_http_variable_t *v;
@@ -1062,7 +520,7 @@ static char* ngx_http_mysql_escape(ngx_conf_t *cf, ngx_command_t *cmd, void *con
 	return NGX_CONF_OK;
 }
 
-static ngx_int_t ngx_http_mysql_init(ngx_conf_t *cf) 
+ngx_int_t ngx_http_mysql_init(ngx_conf_t *cf) 
 {
 	ngx_http_core_main_conf_t *cmcf;
 	ngx_http_handler_pt *h;
